fix(geometry): the seam radius was rounded to float between two doubles
`Surface::radius` is a double, `Curve::radius` is a double, and every branch
of intersectSurfaces computes its result as a double — sqrt(R²−d²) for a
plane section, the chord formula for sphere against sphere. The one
file-local helper that assembled the Curve took its radius as a FLOAT, so
the value was rounded on the way between two double quantities. One word.

This is the same failure mode the double migration ended on and the same
place it hides: not in a declaration, but in a three-line helper's
parameter, where the type that matters is neither the input's nor the
output's. BRepSurfaceIntersect had already had its `dot` widened for
exactly this reason; `circleCurve` was missed.

MEASURED on box(2³) against sphere(r=1.2): the seam came back with radius
0.66332507133483887 where the exact value is 0.66332504433416373 — out by
2.7e-08, float resolution at unit scale and nothing to do with the
geometry. What makes it bite is that the two operands do not both inherit
it: the box's ring is cut AT the reported radius while the sphere's ring is
pinned to the sphere and the cutting plane and lands at the true one, so a
single circle becomes two concentric rings 2.7e-08 apart. Widened, they
agree to 5.6e-17.

test_BRepSeamRadiusPrecision.cpp (4 tests), asserting RELATIVE agreement at
scale 1, 1e3 and 1e6. Both halves matter: a float in the chain is a fixed
relative error, so a single-scale absolute bound can be satisfied by
shrinking the model instead of fixing the arithmetic — which is how this
survived the migration. Load-bearing by revert: the float misses 1e-15 by
seven orders, and its relative error GROWS with scale (3.1e-09 → 3.5e-08),
which is the signature itself. Tangency still reports Point, not a
zero-radius Circle.

SCOPE, stated plainly: this is a real narrowing and it is fixed, but it is
NOT why box/sphere sews to empty. That failure survives the fix unchanged,
so the comment in the source says so rather than claiming a cure.

Suite 2512 ran / 2507 pass / 5 hardware skips, zero regressions.

+++ apps/Nexus-Modeling/src/kernel/src/geometry/BRepSurfaceIntersect.cpp
@@ -52,7 +52,23 @@ Curve lineCurve(const Vec3& origin, const Vec3& dir)
     c.dir = normalize(dir);
     return c;
 }
-Curve circleCurve(const Vec3& center, const Vec3& axis, float radius)
+// DOUBLE radius, for the same reason `dot` above returns double. Every caller computes
+// the radius as a double — sqrt(R² − d²) for a plane section, the chord formula for
+// sphere∩sphere — and `Curve::radius` stores a double, so a float parameter here rounded
+// the value on its way between two double quantities and nothing downstream could see it.
+//
+// Measured on box(2³) against sphere(r=1.2): the seam radius came out
+// 0.66332507133483887 where the exact value is 0.66332504433416373 — wrong by 2.7e-08,
+// which is float resolution at this magnitude and nothing to do with the geometry. The
+// box's ring was cut at the rounded radius while the sphere's ring, whose vertices are
+// pinned to both the sphere and the cutting plane, landed at the exact one, leaving two
+// concentric rings 2.7e-08 apart where there is only one circle. With the parameter
+// widened they agree to 5.6e-17.
+//
+// Scope, stated honestly: this is a real narrowing and it is fixed here, but it is NOT
+// why box/sphere sews to empty. That failure survives the fix unchanged, because the two
+// rings are duplicated as SEPARATE VERTICES regardless of how exactly they agree.
+Curve circleCurve(const Vec3& center, const Vec3& axis, double radius)
 {
     Curve c;
     c.kind = CurveKind::Circle;
